# rest/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Paperless REST API
  description: |
    REST API for the Paperless Document Management System.
    This API provides CRUD operations for managing file metadata including
    upload time, filename, author, last edited, size, and file type.
  version: 1.0.0
  contact:
    name: Paperless Document Management Team
    email: support@paperless.com
servers:
  - url: http://localhost:8081
    description: Local development server

tags:
  - name: File Metadata Management
    description: API for managing file metadata in the paperless document system

paths:
  /api/v1/files:
    post:
      tags:
        - File Metadata Management
      summary: Upload file with metadata
      description: Upload a file and create metadata entry
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - author
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
                author:
                  type: string
                  description: Author of the document
                  example: "John Doe"
      responses:
        '201':
          description: File metadata created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetadataResponseDto'
              examples:
                example1:
                  summary: Created PDF Document
                  value:
                    id: 1
                    filename: "document.pdf"
                    author: "John Doe"
                    fileType: "PDF"
                    size: 1024000
                    uploadTime: "2025-10-08T15:30:00"
                    lastEdited: "2025-10-08T15:30:00"
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validationError:
                  summary: Validation Error
                  value:
                    timestamp: "2025-10-08T15:30:00"
                    status: 400
                    error: "Bad Request"
                    message: "Filename is required"
                    path: "/api/v1/files"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                serverError:
                  summary: Internal Server Error
                  value:
                    timestamp: "2025-10-08T15:30:00"
                    status: 500
                    error: "Internal Server Error"
                    message: "An unexpected error occurred"
                    path: "/api/v1/files"

    get:
      tags:
        - File Metadata Management
      summary: Get all file metadata
      description: Retrieve all file metadata entries, optionally filtered
      operationId: getAllFileMetadata
      parameters:
        - name: search
          in: query
          description: Search keyword (optional)
          required: false
          schema:
            type: string
          example: "document"
        - name: author
          in: query
          description: Filter by author (optional)
          required: false
          schema:
            type: string
          example: "John Doe"
        - name: fileType
          in: query
          description: Filter by file type (optional)
          required: false
          schema:
            type: string
          example: "PDF"
      responses:
        '200':
          description: File metadata retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileMetadataResponseDto'
              examples:
                multipleFiles:
                  summary: Multiple Files
                  value:
                    - id: 1
                      filename: "document.pdf"
                      author: "John Doe"
                      fileType: "PDF"
                      size: 1024000
                      uploadTime: "2025-10-08T15:30:00"
                      lastEdited: "2025-10-08T15:30:00"
                    - id: 2
                      filename: "report.docx"
                      author: "Jane Smith"
                      fileType: "DOCX"
                      size: 512000
                      uploadTime: "2025-10-08T16:00:00"
                      lastEdited: "2025-10-08T16:00:00"
                emptyList:
                  summary: No Files
                  value: []
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                serverError:
                  summary: Internal Server Error
                  value:
                    timestamp: "2025-10-08T15:30:00"
                    status: 500
                    error: "Internal Server Error"
                    message: "An unexpected error occurred"
                    path: "/api/v1/files"

  /api/v1/files/{id}:
    get:
      tags:
        - File Metadata Management
      summary: Get file metadata by ID
      description: Retrieve file metadata by its unique identifier
      operationId: getFileMetadata
      parameters:
        - name: id
          in: path
          description: File metadata ID
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: File metadata retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetadataResponseDto'
              examples:
                singleFile:
                  summary: Single File
                  value:
                    id: 1
                    filename: "document.pdf"
                    author: "John Doe"
                    fileType: "PDF"
                    size: 1024000
                    uploadTime: "2025-10-08T15:30:00"
                    lastEdited: "2025-10-08T15:30:00"
        '404':
          description: File metadata not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  summary: File Not Found
                  value:
                    timestamp: "2025-10-08T15:30:00"
                    status: 404
                    error: "Not Found"
                    message: "FileMetadata with id 999 not found"
                    path: "/api/v1/files/999"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                serverError:
                  summary: Internal Server Error
                  value:
                    timestamp: "2025-10-08T15:30:00"
                    status: 500
                    error: "Internal Server Error"
                    message: "An unexpected error occurred"
                    path: "/api/v1/files/999"

    patch:
      tags:
        - File Metadata Management
      summary: Update file metadata
      description: Partially update file metadata and optionally replace the file
      operationId: updateFileMetadata
      parameters:
        - name: id
          in: path
          description: File metadata ID
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: New file to replace existing one (optional)
                author:
                  type: string
                  description: Author of the document (optional)
                  example: "John Smith"
      responses:
        '200':
          description: File metadata updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetadataResponseDto'
              examples:
                updatedFile:
                  summary: Updated File
                  value:
                    id: 1
                    filename: "updated_document.pdf"
                    author: "John Smith"
                    fileType: "PDF"
                    size: 1048576
                    uploadTime: "2025-10-08T15:30:00"
                    lastEdited: "2025-10-08T17:00:00"
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validationError:
                  summary: Validation Error
                  value:
                    timestamp: "2025-10-08T15:30:00"
                    status: 400
                    error: "Bad Request"
                    message: "Invalid input data"
                    path: "/api/v1/files/1"
        '404':
          description: File metadata not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  summary: File Not Found
                  value:
                    timestamp: "2025-10-08T15:30:00"
                    status: 404
                    error: "Not Found"
                    message: "FileMetadata with id 999 not found"
                    path: "/api/v1/files/999"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                serverError:
                  summary: Internal Server Error
                  value:
                    timestamp: "2025-10-08T15:30:00"
                    status: 500
                    error: "Internal Server Error"
                    message: "An unexpected error occurred"
                    path: "/api/v1/files/999"

    delete:
      tags:
        - File Metadata Management
      summary: Delete file metadata
      description: Delete file metadata by its unique identifier
      operationId: deleteFileMetadata
      parameters:
        - name: id
          in: path
          description: File metadata ID
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          description: File metadata deleted successfully
        '404':
          description: File metadata not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  summary: File Not Found
                  value:
                    timestamp: "2025-10-08T15:30:00"
                    status: 404
                    error: "Not Found"
                    message: "FileMetadata with id 999 not found"
                    path: "/api/v1/files/999"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                serverError:
                  summary: Internal Server Error
                  value:
                    timestamp: "2025-10-08T15:30:00"
                    status: 500
                    error: "Internal Server Error"
                    message: "An unexpected error occurred"
                    path: "/api/v1/files/999"

components:
  schemas:
    FileMetadataResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the file metadata
          example: 1
        filename:
          type: string
          description: Name of the file
          example: "document.pdf"
        author:
          type: string
          description: Author of the file
          example: "John Doe"
        fileType:
          type: string
          description: Type/format of the file
          example: "PDF"
        size:
          type: integer
          format: int64
          description: Size of the file in bytes
          example: 1024000
        uploadTime:
          type: string
          format: date-time
          description: Timestamp when the file was uploaded
          example: "2025-10-08T15:30:00"
        lastEdited:
          type: string
          format: date-time
          description: Timestamp when the file metadata was last modified
          example: "2025-10-08T15:30:00"
      example:
        id: 1
        filename: "document.pdf"
        author: "John Doe"
        fileType: "PDF"
        size: 1024000
        uploadTime: "2025-10-08T15:30:00"
        lastEdited: "2025-10-08T15:30:00"

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
          example: "2025-10-08T15:30:00"
        status:
          type: integer
          description: HTTP status code
          example: 404
        error:
          type: string
          description: Error type
          example: "Not Found"
        message:
          type: string
          description: Error message
          example: "FileMetadata with id 999 not found"
        path:
          type: string
          description: Request path that caused the error
          example: "/api/v1/files/999"
      example:
        timestamp: "2025-10-08T15:30:00"
        status: 404
        error: "Not Found"
        message: "FileMetadata with id 999 not found"
        path: "/api/v1/files/999"